TST: Add test for _verify_variant_name, move to _db, remove pandas
