+++ mitosis/_db.py
@@ -16,6 +16,7 @@ from sqlalchemy import Table
 from sqlalchemy import Update
 from sqlalchemy import update
 
+from mitosis._strings import cleanstr
 from mitosis._typing import Parameter
 
 
@@ -116,3 +117,33 @@ def _id_variant_iteration(eng: Engine, trials_table: Table, master_variant: str)
         return 0
     else:
         return max(row[0] for row in rows) + 1
+
+
+def _verify_variant_name(trial_db: Path, step: str, param: Parameter) -> None:
+    """Check for conflicts between variant names in prior trials
+
+    Side effects:
+        - If trial_db does not exist, will create it
+        - If variant name has not been used before, will insert it
+    """
+    eng = create_engine("sqlite:///" + str(trial_db))
+    md = MetaData()
+    tb = Table(f"{step}_variant_{param.arg_name}", md, *variant_types())
+    try:
+        val_str = cleanstr(param.vals)
+    except Exception:
+        raise RuntimeError(f"Unable to establish reproducible {param=}")
+    with eng.connect() as conn:
+        names_and_vals = conn.execute(select(tb.c.name, tb.c.params))
+    vals_matching_name = [val for name, val in names_and_vals if name == param.var_name]
+    if len(vals_matching_name) == 0:
+        stmt = tb.insert().values({"name": param.var_name, "params": val_str})
+        with eng.begin() as conn:
+            conn.execute(stmt)
+    elif vals_matching_name[0][0] != val_str:
+        raise RuntimeError(
+            f"Parameter '{param.arg_name}' variant '{param.var_name}' "
+            f"is stored with different values in {trial_db}, table '{tb}'. "
+            f"(Stored: {vals_matching_name[0][0]}), attmpeted: {val_str}."
+        )
+    # Otherwise, parameter has already been registered and no conflicts
