+++ pyproject.toml
@@ -42,7 +42,6 @@ dependencies = [
   "nbconvert",
   "nbclient",
   "nbformat",
-  "pandas<2.2",
   "sqlalchemy>=2.0",
   "toml",
   "types-toml",
@@ -57,7 +56,6 @@ dev = [
   "black",
   "coverage",
   "mypy",
-  "pandas-stubs",
   "pre-commit",
   "pytest-lazy-fixture",
   "sphinx",
